fix test imports for renamed triron-test
