+++ shared/ui-components/src/headers/Hero/Hero.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { withA11y } from "@storybook/addon-a11y"
 import Hero from "./Hero"
 import Archer from "@bloom-housing/listings-service/listings/archer.json"
-import Triton from "@bloom-housing/listings-service/listings/triton.json"
+import Triton from "@bloom-housing/listings-service/listings/triton-test.json"
 import { Listing } from "@bloom-housing/core"
 
 export default {
